Skip maintaining index for shard_by_attr (#77)
Maintaining index sets for shard_by_attr is unnecessary. It's much more performant if we don't have to join it with other indices sets (especially when the other sets are ranged index sets -- we'd perform a loop join in the lua script).

This PR addresses the issue and enforces queries to contain at least 2 query conditions: the shard by attr condition, and another query condition on an indexed attribute (or more conditions).

For custom_index, the first attr (previously the shard_by_attr) is now removed -- the data in the zset for the first attr is the same for all records in the set anyway.